Integrate FDataBasis with Pandas
