+++ skfda/representation/grid.py
@@ -24,7 +24,7 @@ __author__ = "Miguel Carbajo Berrocal"
 __email__ = "miguel.carbajo@estudiante.uam.es"
 
 
-class FDataGrid(FData, pandas.api.extensions.ExtensionArray):
+class FDataGrid(FData):
     r"""Represent discretised functional data.
 
     Class for representing functional data as a set of curves discretised
@@ -983,6 +983,10 @@ class FDataGrid(FData, pandas.api.extensions.ExtensionArray):
         else:
             return self.copy(data_matrix=self.data_matrix[key])
 
+    #####################################################################
+    # Numpy methods
+    #####################################################################
+
     def __array_ufunc__(self, ufunc, method, *inputs, **kwargs):
 
         for i in inputs:
@@ -1025,26 +1029,6 @@ class FDataGrid(FData, pandas.api.extensions.ExtensionArray):
         return FDataGridDType
 
     @property
-    def ndim(self):
-        """
-        Return number of dimensions of the functional data. It is
-        always 1, as each observation is considered a "scalar" object.
-
-        Returns:
-            int: Number of dimensions of the functional data.
-
-        """
-        return 1
-
-    @classmethod
-    def _from_sequence(cls, scalars, dtype=None, copy=False):
-        return cls(scalars, dtype=dtype)
-
-    @classmethod
-    def _from_factorized(cls, values, original):
-        return cls(values)
-
-    @property
     def nbytes(self) -> int:
         """
         The number of bytes needed to store this object in memory.
@@ -1052,99 +1036,14 @@ class FDataGrid(FData, pandas.api.extensions.ExtensionArray):
         return self.data_matrix.nbytes() + sum(
             p.nbytes() for p in self.sample_points)
 
-    def isna(self):
-        """
-        A 1-D array indicating if each value is missing.
-
-        Returns:
-            na_values (np.ndarray): Array full of True values.
-        """
-        return np.ones(self.nsamples, dtype=bool)
-
-    def take(self, indices, allow_fill=False, fill_value=None):
-        """
-        Take elements from an array.
-        Parameters:
-            indices (sequence of integers):
-                Indices to be taken.
-            allow_fill (bool, default False): How to handle negative values
-                in `indices`.
-                * False: negative values in `indices` indicate positional
-                  indices from the right (the default). This is similar to
-                  :func:`numpy.take`.
-                * True: negative values in `indices` indicate
-                  missing values. These values are set to `fill_value`. Any
-                  other negative values raise a ``ValueError``.
-            fill_value (any, optional):
-                Fill value to use for NA-indices when `allow_fill` is True.
-                This may be ``None``, in which case the default NA value for
-                the type, ``self.dtype.na_value``, is used.
-                For many ExtensionArrays, there will be two representations of
-                `fill_value`: a user-facing "boxed" scalar, and a low-level
-                physical NA value. `fill_value` should be the user-facing
-                version, and the implementation should handle translating that
-                to the physical version for processing the take if necessary.
-        Returns:
-            FDataGrid
-        Raises:
-            IndexError: When the indices are out of bounds for the array.
-            ValueError: When `indices` contains negative values other than
-                        ``-1`` and `allow_fill` is True.
-        Notes:
-            ExtensionArray.take is called by ``Series.__getitem__``, ``.loc``,
-            ``iloc``, when `indices` is a sequence of values. Additionally,
-            it's called by :meth:`Series.reindex`, or any other method
-            that causes realignment, with a `fill_value`.
-        See Also:
-            numpy.take
-            pandas.api.extensions.take
-        """
-        from pandas.core.algorithms import take
-        # If the ExtensionArray is backed by an ndarray, then
-        # just pass that here instead of coercing to object.
-        data = self.astype(object)
-        if allow_fill and fill_value is None:
-            fill_value = self.dtype.na_value
-        # fill value should always be translated from the scalar
-        # type for the array, to the physical storage type for
-        # the data, before passing to take.
-        result = take(data, indices, fill_value=fill_value,
-                      allow_fill=allow_fill)
-        return self._from_sequence(result, dtype=self.dtype)
-
-    @classmethod
-    def _concat_same_type(
-            cls,
-            to_concat
-    ):
-        """
-        Concatenate multiple array
-
-        Parameters:
-            to_concat (sequence of FDataGrid)
-        Returns:
-            FDataGrid
-        """
-
-        first, *others = to_concat
-
-        for o in others:
-            first = first.concatenate(o)
-
-        # When #101 is ready
-        # return first.concatenate(others)
-
-        return first
-
 
 class FDataGridDType(pandas.api.extensions.ExtensionDtype):
     """
     DType corresponding to FDataGrid in Pandas
     """
-    name = 'discretized functional data (grid)'
+    name = 'functional data (grid)'
     kind = 'O'
     type = FDataGrid
-    _record_type = np.dtype([('hi', '>u8'), ('lo', '>u8')])
     na_value = None
 
     @classmethod
